// src/spiders/broker/broker.interface.ts
export interface Broker {
  /**
   * 编码
   */
  code: string
  /**
   * 真实姓名
   */
  name: string
  /**
   * 性别
   */
  gender: string
  /**
   * 联系电话
   */
  tel: string
  /**
   * 从业机构代码
   */
  orgCode: string
  /**
   * 从业机构名称
   */
  orgName: string
  /**
   * 签约状态
   */
  status: string
}